configure.ac: Re-work macro initialisation
Seems the current automake/autoconf aren't happy with the macros in this
file, or I'm just not that adept at setting up autoconf/automake
properly.

(In fact, the latter is a distinct possibility.)

The following is based on what I've seen from the following links
(monkey see, monkey do):
- http://stackoverflow.com/questions/16188335/automake-error-no-proper-invocation-of-am-init-automake-was-found
- http://www.gnu.org/software/automake/manual/automake.html#Hello-World
- http://www.gnu.org/software/autoconf/manual/autoconf.html#AC_005fINIT

This also adds a bug report email address and link to the homepage to
the generated ./configure script.

These changes work with autoconf 2.69, automake 1.15.  To generate the
configure scripts, run:

	$ autoreconf --install